move utilities to packages dir
