+++ mealie/pkgs/cache/cache_key.py
@@ -2,8 +2,7 @@ import random
 import string
 
 
-def new_cache_key(length=4) -> str:
+def new_key(length=4) -> str:
     """returns a 4 character string to be used as a cache key for frontend data"""
     options = string.ascii_letters + string.digits
-
     return "".join(random.choices(options, k=length))
