+++ mealie/pkgs/stats/__init__.py
@@ -1,0 +1,1 @@
+from .fs_stats import *
