+++ mealie/pkgs/stats/fs_stats.py
