+++ mealie/routes/backup_routes.py
@@ -16,7 +16,7 @@ from mealie.schema.user.user import PrivateUser
 from mealie.services.backups import imports
 from mealie.services.backups.exports import backup_all
 from mealie.services.events import create_backup_event
-from mealie.utils.fs_stats import pretty_size
+from mealie.pkgs.stats.fs_stats import pretty_size
 
 router = AdminAPIRouter(prefix="/api/backups", tags=["Backups"])
 logger = get_logger()
