--- mealie/utils/__init__.py
@@ -1,1 +1,0 @@
-from .cache_key import new_cache_key
